requirements.txt
Django==4.1.11
djangorestframework==3.12.4
django-rest-knox==4.2.0
requests==2.25.1
drf-yasg==1.20.0
pandas==1.3.3

